pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "torino"
version = "0.1.0"
description = "Multi-agentic JIRA triage tool for Red Hat Satellite"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "jira>=3.8",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
torino = "torino.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
